fix: normalize subject legacy scope keys

+++ EduOS.Core/Entities/Academic/Subject.cs
@@ -7,8 +7,8 @@ namespace EduOS.Core.Entities.Academic
     public class Subject : BaseTenantEntity
     {
         // Temporary compatibility scope while legacy Class/Group workflows migrate to CurriculumSubject.
-        public int ClassId { get; set; }
-        public int? GroupId { get; set; }
+        public long ClassId { get; set; }
+        public long? GroupId { get; set; }
 
         [Required, MaxLength(200)]
         public string Name { get; set; } = string.Empty;
